Refactor torch_runner in PytorchEstimator for ray and pyspark backend (#3541)
* refactor torch runner

* rename state_stream -> get_state_stream

* continue refactor runnner

* rename pyspark runner to pyspark worker

* fix ray tests

* fix pyspark

* torch_pyspark_runner -> pytorch_pyspark_worker

* fix horovod test

* add shutdown
