+++ python/orca/src/bigdl/orca/learn/pytorch/pytorch_pyspark_estimator.py
@@ -21,7 +21,7 @@ import torch
 import numpy as np
 
 from bigdl.orca.learn.pytorch.training_operator import TrainingOperator
-from bigdl.orca.learn.pytorch.torch_pyspark_runner import TorchPysparkRunner
+from bigdl.orca.learn.pytorch.pytorch_pyspark_worker import PytorchPysparkWorker
 from bigdl.orca.learn.utils import maybe_dataframe_to_xshards, dataframe_to_xshards, \
     convert_predict_xshards_to_dataframe, make_data_creator, update_predict_xshards, \
     process_xshards_of_pandas_dataframe
@@ -30,7 +30,7 @@ from bigdl.orca import OrcaContext
 from bigdl.orca.learn.base_estimator import BaseEstimator
 from bigdl.dllib.utils.file_utils import enable_multi_fs_load, enable_multi_fs_save
 from bigdl.dllib.utils.common import get_node_and_core_number
-from bigdl.orca.learn.pytorch.torch_pyspark_runner import find_ip_and_port
+from bigdl.orca.learn.pytorch.pytorch_pyspark_worker import find_ip_and_port
 
 
 logger = logging.getLogger(__name__)
@@ -127,7 +127,7 @@ class PyTorchPySparkEstimator(BaseEstimator):
             cluster_info=self._get_cluster_info(sc)
         )
 
-        self.driver_runner = TorchPysparkRunner(**self.worker_init_params, mode='predict')
+        self.driver_runner = PytorchPysparkWorker(**self.worker_init_params, mode='predict')
 
         self.state_dict = self.driver_runner.get_state_dict()
 
@@ -199,7 +199,7 @@ class PyTorchPySparkEstimator(BaseEstimator):
             def transform_func(iter, init_params, param):
                 partition_data = list(iter)
                 param["data_creator"] = partition_to_creator(partition_data)
-                runner = TorchPysparkRunner(**init_params)
+                runner = PytorchPysparkWorker(**init_params)
                 result = runner.train_epochs(**param)
                 runner.shutdown()
                 return result
@@ -216,7 +216,7 @@ class PyTorchPySparkEstimator(BaseEstimator):
             params["data_creator"] = data
 
             def transform_func(iter, init_param, param):
-                return TorchPysparkRunner(**init_param).train_epochs(**param)
+                return PytorchPysparkWorker(**init_param).train_epochs(**param)
 
             res = self.workerRDD.barrier().mapPartitions(
                 lambda iter: transform_func(iter, init_params, params)).collect()
@@ -230,7 +230,7 @@ class PyTorchPySparkEstimator(BaseEstimator):
             partition_data = list(iter)
             # res = combine_in_partition(partition_data)
             param["data_creator"] = make_data_creator(partition_data)
-            return TorchPysparkRunner(**init_param).predict(**params)
+            return PytorchPysparkWorker(**init_param).predict(**params)
 
         pred_shards = SparkXShards(xshards.rdd.mapPartitions(
                                         lambda iter: transform_func(iter, init_params, params)))
@@ -341,7 +341,7 @@ class PyTorchPySparkEstimator(BaseEstimator):
             def transform_func(iter, init_param, param):
                 partition_data = list(iter)
                 param["data_creator"] = partition_to_creator(partition_data)
-                return TorchPysparkRunner(**init_param).validate(**param)
+                return PytorchPysparkWorker(**init_param).validate(**param)
 
             res = data.rdd.repartition(self.num_workers).barrier() \
                 .mapPartitions(lambda iter: transform_func(iter, init_params, params)).collect()
@@ -349,7 +349,7 @@ class PyTorchPySparkEstimator(BaseEstimator):
             params["data_creator"] = data
 
             def transform_func(iter, init_param, param):
-                return TorchPysparkRunner(**init_param).validate(**param)
+                return PytorchPysparkWorker(**init_param).validate(**param)
 
             res = self.workerRDD.barrier().mapPartitions(
                 lambda iter: transform_func(iter, init_params, params)).collect()
