+++ python/orca/test/bigdl/orca/learn/ray/pytorch/test_estimator_pytorch_backend.py
@@ -165,7 +165,7 @@ class TestPyTorchEstimator(TestCase):
         # Verify syncing weights, i.e. the two workers have the same weights after training
         import ray
         remote_workers = estimator.remote_workers
-        state_dicts = ray.get([worker.state_dict.remote() for worker in remote_workers])
+        state_dicts = ray.get([worker.get_state_dict.remote() for worker in remote_workers])
         weights = [state["models"] for state in state_dicts]
         worker1_weights = weights[0][0]
         worker2_weights = weights[1][0]
